pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "4.11.2"
requests = "2.28.2"
selenium = "4.8.2"
python-dateutil = "2.8.2"
progress = "1.6"
feedparser = "6.0.10"
google-api-python-client = "2.80.0"
google-auth-httplib2 = "0.1.0"
google-auth-oauthlib = "1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
